+++ apps/caelus/source/commands/chat-inputs/schedule.ts
@@ -1,6 +1,6 @@
 import { type APIChatInputApplicationCommandInteraction, Locale } from "@discordjs/core";
 import { t } from "i18next";
-import { scheduleOverview } from "../../services/schedule.js";
+import { scheduleOverview } from "../../features/schedule.js";
 
 export default {
 	name: t("schedule.command-name", { lng: Locale.EnglishGB, ns: "commands" }),
